+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased (0.5.0)
+### Added
+- Average for the last 28 days script.
+
 ## 0.4.0
 ### Added
 - Script to generate test data for reporting tests and graphing.